# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Training Data API",
    "version": "1.0.0",
    "description": "Standalone compatibility API for the grocery delivery training dataset. Read routes are public; quote and order routes require x-api-key or apikey."
  },
  "servers": [{ "url": "https://train.uxi.asia" }],
  "components": {
    "securitySchemes": {
      "apiKey": { "type": "apiKey", "in": "header", "name": "x-api-key" }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["code", "message"],
        "properties": { "code": { "type": "string" }, "message": { "type": "string" } }
      }
    }
  },
  "paths": {
...
    "/rest/v1/grocery_catalog": {
      "get": {
        "summary": "List active grocery catalog items",
        "parameters": [
          { "name": "select", "in": "query", "schema": { "type": "string" }, "description": "Comma-separated supported fields" },
          { "name": "is_active", "in": "query", "schema": { "type": "string", "enum": ["eq.true", "eq.false"] } }
        ],
        "responses": { "200": { "description": "Array of catalog rows" }, "400": { "description": "Invalid query", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } } }
      }
    },
...
    "/rest/v1/rpc/create_order_quote": {
      "post": {
        "summary": "Create a short-lived price quote",
        "security": [{ "apiKey": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["p_item_id", "p_quantity"], "properties": { "p_item_id": { "type": "string", "format": "uuid" }, "p_quantity": { "type": "integer", "minimum": 1, "maximum": 100 } } } } } },
        "responses": { "200": { "description": "Single-row quote array" }, "400": { "description": "Invalid request" }, "401": { "description": "API key required" }, "404": { "description": "Item not found" }, "409": { "description": "Insufficient inventory" } }
      }
    },
...
  }
}